backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
python-multipart
redis
rq
openai
cryptography
pypdf
fpdf2
sentence-transformers
